Split components into seperate files
