+++ bashable_notes_client/src/uikit.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const Container = ({children}) => (
+	<div className="uk-container">{children}</div>
+);
+
+export const Spinner = ({}) => (
+	<div uk-spinner=""></div>
+);
